refactor: 픽잇 결과 데이터 요청  로딩 상태 및 에러 처리
